moved tools and utils out of common
