+++ src/jadhav_lab_to_nwb/rivera_and_shukla_2025/rivera_and_shukla_2025_dlc_interface.py
@@ -1,7 +1,7 @@
 """Primary class for converting experiment-specific behavioral video."""
 from neuroconv.utils import DeepDict, dict_deep_update
 
-from ..common.utils.utils import rivera_and_shukla_2025_get_epoch_name
+from ..utils.utils import rivera_and_shukla_2025_get_epoch_name
 from ..datainterfaces.base_dlc_interface import BaseDeepLabCutInterface
