+++ src/jadhav_lab_to_nwb/rivera_and_shukla_2025/rivera_and_shukla_2025_epoch_interface.py
@@ -1,7 +1,7 @@
 """Primary class for converting experiment-specific behavior."""
 from pynwb.file import NWBFile
 
-from ..common.utils.utils import rivera_and_shukla_2025_get_epoch_name
+from ..utils.utils import rivera_and_shukla_2025_get_epoch_name
 from ..datainterfaces.base_epoch_interface import BaseEpochInterface
