+++ src/jadhav_lab_to_nwb/rivera_and_shukla_2025/rivera_and_shukla_2025_nwbconverter.py
@@ -8,7 +8,7 @@ from jadhav_lab_to_nwb.rivera_and_shukla_2025 import (
     RiveraAndShukla2025EpochInterface,
 )
 
-from ..common.tools.spikegadgets import readCameraModuleTimeStamps
+from ..tools.spikegadgets import readCameraModuleTimeStamps
 
 
 class RiveraAndShukla2025NWBConverter(NWBConverter):
